pyproject.toml
[project]
name = "market"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "django-silk>=5.3.2",
    "pre-commit>=4.1.0",
    "ruff==0.9.6",
]
